+++ package.json
@@ -1,5 +1,5 @@
 {
-  "dependencies": {
-    "serverless-python-requirements": "^6.1.0"
-  }
+    "dependencies": {
+      "serverless-python-requirements": "^6.1.0"
+    }
 }
