+++ requirements.txt
@@ -10,4 +10,4 @@ uvicorn==0.17.1
 pydantic==1.9.0
 PyJWT==2.3.0
 email-validator==1.3.0
-pytest==6.2.5
+httpx~=0.27.2
